devtools/client/shared/components/reps/test/mochitest/test_reps_attribute.html

<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
<!DOCTYPE HTML>
<html>
<!--
Test Attribute rep
-->
<head>
  <meta charset="utf-8">
  <title>Rep test - Attribute</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="head.js" type="application/javascript"></script>
<script type="application/javascript">
window.onload = Task.async(function* () {
  const {
    REPS,
    getRep,
  } = browserRequire("devtools/client/shared/components/reps/reps");
  let { Attribute } = REPS;

  try {
    testBasic();
    testObjectLink();
  } catch (e) {
    ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
  } finally {
    SimpleTest.finish();
  }

  function testBasic() {
    // Test that correct rep is chosen
    is(getRep(getStub()), Attribute.rep, "Rep correctly selects Attribute Rep");

    // Test rendering
    const renderedComponent = renderComponent(Attribute.rep, { object: getStub() });
    is(renderedComponent.textContent, "class=\"autocomplete-suggestions\"", "Attribute rep has expected text content");
  }

  function testObjectLink() {
    const renderedComponent = renderComponent(Attribute.rep, {
      object: getStub(),
      objectLink: (props, ...children) => React.DOM.span({},
        "*", ...children, "*"),
    });
    is(renderedComponent.textContent, "*class=\"autocomplete-suggestions\"*",
      "Attribute rep has expected text content when an objectLink is passed as a prop");
  }

  function getStub() {
    return {
      "type": "object",
      "class": "Attr",
      "actor": "server1.conn19.obj65",
      "extensible": true,
      "frozen": false,
      "sealed": false,
      "ownPropertyLength": 0,
      "preview": {
        "kind": "DOMNode",
        "nodeType": 2,
        "nodeName": "class",
        "value": "autocomplete-suggestions"
      }
    };
  }
});
</script>
</pre>
</body>
</html>
